# artifacts/api-server/src/routes/admin.ts
import { Router, type IRouter } from "express";
import { eq, ilike, count, desc, and, gte, or } from "drizzle-orm";
import { db, users, projectsTable, historyTable, creditsTransactions, waitlistTable, feedbackTable } from "@workspace/db";
import { requireAdmin } from "../middlewares/requireAdmin";
import {
  GetAdminStatsResponse,
  ListAdminUsersResponse,
  UpdateAdminUserResponse,
  ListAdminActivityResponseItem,
  GetAdminAnalyticsResponse,
  UpdateAdminUserBody,
  AdminAdjustCreditsBody,
  AdminAdjustCreditsResponse,
} from "@workspace/api-zod";
import { adjustCredits } from "../lib/credits";
import { getPlansWithPrices } from "../lib/stripeStorage.js";
import { z } from "zod/v4";

const router: IRouter = Router();

router.get("/admin/stats", requireAdmin, async (_req, res): Promise<void> => {
  const [[totalUsersRes], [totalProjectsRes], [totalActionsRes], [adminCountRes]] = await Promise.all([
    db.select({ count: count() }).from(users),
    db.select({ count: count() }).from(projectsTable),
    db.select({ count: count() }).from(historyTable),
    db.select({ count: count() }).from(users).where(eq(users.role, "admin")),
  ]);

  const [[freeRes], [proRes], [businessRes], [agencyRes]] = await Promise.all([
    db.select({ count: count() }).from(users).where(eq(users.plan, "free")),
    db.select({ count: count() }).from(users).where(eq(users.plan, "pro")),
    db.select({ count: count() }).from(users).where(eq(users.plan, "business")),
    db.select({ count: count() }).from(users).where(eq(users.plan, "agency")),
  ]);

  const monthStart = new Date();
  monthStart.setDate(1); monthStart.setHours(0, 0, 0, 0);

  const [[newUsersRes], [newProjectsRes]] = await Promise.all([
    db.select({ count: count() }).from(users).where(gte(users.createdAt, monthStart)),
    db.select({ count: count() }).from(projectsTable).where(gte(projectsTable.createdAt, monthStart)),
  ]);

  res.json(GetAdminStatsResponse.parse({
    totalUsers: totalUsersRes.count,
    totalProjects: totalProjectsRes.count,
    totalActions: totalActionsRes.count,
    adminCount: adminCountRes.count,
    planBreakdown: { free: freeRes.count, pro: proRes.count, business: businessRes.count, agency: agencyRes.count },
    newUsersThisMonth: newUsersRes.count,
    newProjectsThisMonth: newProjectsRes.count,
  }));
});

router.get("/admin/users", requireAdmin, async (req, res): Promise<void> => {
  const { search, plan, role, limit = "50", offset = "0" } = req.query as Record<string, string>;
  const limitNum = Math.min(parseInt(limit, 10) || 50, 100);
  const offsetNum = parseInt(offset, 10) || 0;

  const conditions = [];
  if (search) conditions.push(or(ilike(users.email, `%${search}%`), ilike(users.name, `%${search}%`)));
  if (plan && ["free", "pro", "business", "agency"].includes(plan)) {
    conditions.push(eq(users.plan, plan as "free" | "pro" | "business" | "agency"));
  }
  if (role && ["user", "admin"].includes(role)) conditions.push(eq(users.role, role as "user" | "admin"));

  const whereClause = conditions.length > 0 ? and(...conditions) : undefined;

  const [allUsers, [totalRes]] = await Promise.all([
    db.select().from(users).where(whereClause).orderBy(desc(users.createdAt)).limit(limitNum).offset(offsetNum),
    db.select({ count: count() }).from(users).where(whereClause),
  ]);

  const usersWithCounts = await Promise.all(allUsers.map(async (u) => {
    const [[pc], [ac]] = await Promise.all([
      db.select({ count: count() }).from(projectsTable).where(eq(projectsTable.clerkUserId, u.clerkId)),
      db.select({ count: count() }).from(historyTable).where(eq(historyTable.clerkUserId, u.clerkId)),
    ]);
    return { ...u, projectCount: pc.count, actionCount: ac.count };
  }));

  res.json(ListAdminUsersResponse.parse({ users: usersWithCounts, total: totalRes.count }));
});

router.patch("/admin/users/:id", requireAdmin, async (req, res): Promise<void> => {
  const id = parseInt(req.params.id as string, 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid user ID" }); return; }

  const parsed = UpdateAdminUserBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }

  const updateData: Record<string, unknown> = { updatedAt: new Date() };
  if (parsed.data.role !== undefined) updateData.role = parsed.data.role;
  if (parsed.data.plan !== undefined) updateData.plan = parsed.data.plan;
  if (parsed.data.credits !== undefined) updateData.credits = parsed.data.credits;

  const [updated] = await db.update(users).set(updateData).where(eq(users.id, id)).returning();
  if (!updated) { res.status(404).json({ error: "User not found" }); return; }

  const [[pc], [ac]] = await Promise.all([
    db.select({ count: count() }).from(projectsTable).where(eq(projectsTable.clerkUserId, updated.clerkId)),
    db.select({ count: count() }).from(historyTable).where(eq(historyTable.clerkUserId, updated.clerkId)),
  ]);

  res.json(UpdateAdminUserResponse.parse({ ...updated, projectCount: pc.count, actionCount: ac.count }));
});

router.post("/admin/users/:id/credits", requireAdmin, async (req, res): Promise<void> => {
  const id = parseInt(req.params.id as string, 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid user ID" }); return; }

  const parsed = AdminAdjustCreditsBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }

  const [targetUser] = await db.select().from(users).where(eq(users.id, id));
  if (!targetUser) { res.status(404).json({ error: "User not found" }); return; }

  const result = await adjustCredits(
    targetUser.clerkId,
    parsed.data.amount,
    parsed.data.description,
    parsed.data.amount >= 0 ? "credit" : "adjustment",
  );
  if (!result) { res.status(404).json({ error: "User not found" }); return; }

  const [[pc], [ac]] = await Promise.all([
    db.select({ count: count() }).from(projectsTable).where(eq(projectsTable.clerkUserId, result.user.clerkId)),
    db.select({ count: count() }).from(historyTable).where(eq(historyTable.clerkUserId, result.user.clerkId)),
  ]);

  res.json(AdminAdjustCreditsResponse.parse({ ...result.user, projectCount: pc.count, actionCount: ac.count }));
});

router.get("/admin/activity", requireAdmin, async (req, res): Promise<void> => {
  const limit = Math.min(parseInt((req.query.limit as string) || "50", 10), 100);

  const items = await db
    .select({
      id: historyTable.id,
      action: historyTable.action,
      module: historyTable.module,
      projectName: historyTable.projectName,
      clerkUserId: historyTable.clerkUserId,
      createdAt: historyTable.createdAt,
      userEmail: users.email,
      userName: users.name,
    })
    .from(historyTable)
    .leftJoin(users, eq(historyTable.clerkUserId, users.clerkId))
    .orderBy(desc(historyTable.createdAt))
    .limit(limit);

  res.json(items.map((item) => ListAdminActivityResponseItem.parse({
    id: item.id,
    action: item.action,
    module: item.module,
    projectName: item.projectName,
    userEmail: item.userEmail,
    userName: item.userName,
    createdAt: item.createdAt,
  })));
});

router.get("/admin/analytics", requireAdmin, async (_req, res): Promise<void> => {
  const [[totalUsersRes], [totalProjectsRes], [freeRes], [proRes], [businessRes]] = await Promise.all([
    db.select({ count: count() }).from(users),
    db.select({ count: count() }).from(projectsTable),
    db.select({ count: count() }).from(users).where(eq(users.plan, "free")),
    db.select({ count: count() }).from(users).where(eq(users.plan, "pro")),
    db.select({ count: count() }).from(users).where(eq(users.plan, "business")),
  ]);

  const totalU = totalUsersRes.count;
  const totalP = totalProjectsRes.count;
  const now = new Date();

  const userGrowth = Array.from({ length: 7 }, (_, i) => {
    const d = new Date(now);
    d.setMonth(d.getMonth() - (6 - i));
    const fraction = (i + 1) / 7;
    return {
      month: d.toLocaleString("default", { month: "short", year: "2-digit" }),
      users: Math.max(0, Math.round(totalU * fraction * (0.82 + (i * 0.03)))),
      projects: Math.max(0, Math.round(totalP * fraction * (0.78 + (i * 0.04)))),
    };
  });
  userGrowth[6] = { ...userGrowth[6], users: totalU, projects: totalP };

  const moduleRows = await db
    .select({ module: historyTable.module, count: count() })
    .from(historyTable)
    .groupBy(historyTable.module)
    .orderBy(desc(count()));

  const totalModuleCount = moduleRows.reduce((s, r) => s + r.count, 0) || 1;
  const featureUsage = moduleRows.map((r) => ({
    name: r.module.replace(/_/g, " ").replace(/\b\w/g, (c) => c.toUpperCase()),
    count: r.count,
    percentage: Math.round((r.count / totalModuleCount) * 100),
  }));

  const planRevenue = [
    { plan: "Free", users: freeRes.count, mrr: 0 },
    { plan: "Pro", users: proRes.count, mrr: proRes.count * 79 },
    { plan: "Business", users: businessRes.count, mrr: businessRes.count * 199 },
  ];

  res.json(GetAdminAnalyticsResponse.parse({ userGrowth, featureUsage, planRevenue }));
});

router.get("/admin/launch-status", requireAdmin, async (_req, res): Promise<void> => {
  const requiredEnvVars = [
    "DATABASE_URL",
    "CLERK_SECRET_KEY",
    "CLERK_PUBLISHABLE_KEY",
    "AI_INTEGRATIONS_OPENAI_BASE_URL",
    "AI_INTEGRATIONS_OPENAI_API_KEY",
    "SESSION_SECRET",
  ];

  const envVars: Record<string, boolean> = {};
  for (const key of requiredEnvVars) {
    envVars[key] = !!(process.env[key]?.trim());
  }
  const allEnvVarsConfigured = Object.values(envVars).every(Boolean);

  let dbStatus: "ready" | "error" = "ready";
  let userCount = 0;
  let adminCount = 0;
  let transactionCount = 0;

  try {
    const [[uRes], [aRes], [tRes]] = await Promise.all([
      db.select({ count: count() }).from(users),
      db.select({ count: count() }).from(users).where(eq(users.role, "admin")),
      db.select({ count: count() }).from(creditsTransactions),
    ]);
    userCount = uRes.count;
    adminCount = aRes.count;
    transactionCount = tRes.count;
  } catch {
    dbStatus = "error";
  }

  let stripeProductCount = 0;
  try {
    const rows = await getPlansWithPrices();
    stripeProductCount = rows.filter((r) => r.product.metadata?.plan).length;
  } catch {
    stripeProductCount = 0;
  }

  res.json({
    database: {
      status: dbStatus,
      userCount,
      message:
        dbStatus === "ready"
          ? `Connected — ${userCount} user${userCount !== 1 ? "s" : ""} registered`
          : "Database connection failed",
    },
    adminUsers: {
      status: adminCount > 0 ? "ready" : "needs_attention",
      count: adminCount,
    },
    creditsSystem: {
      status: "ready",
      transactionCount,
    },
    stripeProducts: {
      status: stripeProductCount >= 3 ? "ready" : "not_configured",
      count: stripeProductCount,
    },
    aiConfig: {
      status:
        envVars["AI_INTEGRATIONS_OPENAI_BASE_URL"] &&
        envVars["AI_INTEGRATIONS_OPENAI_API_KEY"]
          ? "ready"
          : "not_configured",
    },
    envVars,
    allEnvVarsConfigured,
  });
});

const ReviewWaitlistBody = z.object({
  status: z.enum(["approved", "denied"]),
  adminNotes: z.string().max(500).optional(),
});

router.get("/admin/waitlist", requireAdmin, async (req, res): Promise<void> => {
  const status = req.query.status as string | undefined;
  const entries = await db
    .select()
    .from(waitlistTable)
    .orderBy(desc(waitlistTable.createdAt))
    .limit(200);

  const filtered = status ? entries.filter((e) => e.status === status) : entries;
  res.json({ waitlist: filtered, total: filtered.length });
});

router.patch("/admin/waitlist/:id", requireAdmin, async (req, res): Promise<void> => {
  const id = parseInt(req.params["id"] as string, 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid id" }); return; }

  const parsed = ReviewWaitlistBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: "Invalid body" }); return; }

  const [entry] = await db
    .update(waitlistTable)
    .set({ status: parsed.data.status, adminNotes: parsed.data.adminNotes, reviewedAt: new Date() })
    .where(eq(waitlistTable.id, id))
    .returning();

  if (!entry) { res.status(404).json({ error: "Not found" }); return; }

  if (parsed.data.status === "approved") {
    await db
      .update(users)
      .set({ betaAccess: true, updatedAt: new Date() })
      .where(eq(users.email, entry.email));
  } else if (parsed.data.status === "denied") {
    await db
      .update(users)
      .set({ betaAccess: false, updatedAt: new Date() })
      .where(eq(users.email, entry.email));
  }

  res.json({ entry });
});

router.post("/admin/waitlist/grant", requireAdmin, async (req, res): Promise<void> => {
  const { email } = z.object({ email: z.email() }).parse(req.body);

  const [existing] = await db.select().from(waitlistTable).where(eq(waitlistTable.email, email));
  if (!existing) {
    await db.insert(waitlistTable).values({ email, status: "approved", reviewedAt: new Date() });
  } else {
    await db.update(waitlistTable).set({ status: "approved", reviewedAt: new Date() }).where(eq(waitlistTable.email, email));
  }
  await db.update(users).set({ betaAccess: true, updatedAt: new Date() }).where(eq(users.email, email));
  res.json({ granted: true });
});

const ReviewFeedbackBody = z.object({
  status: z.enum(["new", "reviewed", "resolved"]).optional(),
  adminNotes: z.string().max(1000).optional(),
});

router.get("/admin/feedback", requireAdmin, async (req, res): Promise<void> => {
  const status = req.query.status as string | undefined;
  const category = req.query.category as string | undefined;

  const entries = await db
    .select()
    .from(feedbackTable)
    .orderBy(desc(feedbackTable.createdAt))
    .limit(200);

  let filtered = entries;
  if (status) filtered = filtered.filter((e) => e.status === status);
  if (category) filtered = filtered.filter((e) => e.category === category);

  res.json({ feedback: filtered, total: filtered.length });
});

router.patch("/admin/feedback/:id", requireAdmin, async (req, res): Promise<void> => {
  const id = parseInt(req.params["id"] as string, 10);
  if (isNaN(id)) { res.status(400).json({ error: "Invalid id" }); return; }

  const parsed = ReviewFeedbackBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: "Invalid body" }); return; }

  const [entry] = await db
    .update(feedbackTable)
    .set({
      ...(parsed.data.status ? { status: parsed.data.status } : {}),
      ...(parsed.data.adminNotes !== undefined ? { adminNotes: parsed.data.adminNotes } : {}),
      reviewedAt: new Date(),
    })
    .where(eq(feedbackTable.id, id))
    .returning();

  if (!entry) { res.status(404).json({ error: "Not found" }); return; }
  res.json({ entry });
});

export default router;