# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IAttom Assist API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: plan
          in: query
          schema:
            type: string
            enum: [free, pro, business, agency]
        - name: role
          in: query
          schema:
            type: string
            enum: [user, admin]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update a user's plan, role, or credits
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/credits:
    post:
      operationId: adminAdjustCredits
      tags: [admin]
      summary: Adjust a user's credits with a transaction record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAdjustCreditsBody"
      responses:
        "200":
          description: Adjustment applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/activity:
    get:
      operationId: listAdminActivity
      tags: [admin]
      summary: Platform-wide activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Platform activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminActivityItem"
...
components:
  schemas:
...
    AdminAdjustCreditsBody:
      type: object
      properties:
        amount:
          type: integer
        description:
          type: string
      required:
        - amount
        - description
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        plan:
          type: string
          enum: [free, pro, business, agency]
        credits:
          type: integer
        projectCount:
          type: integer
        actionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - role
        - plan
        - credits
        - projectCount
        - actionCount
        - createdAt
    AdminUserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
      required:
        - users
        - total
    UpdateAdminUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        plan:
          type: string
          enum: [free, pro, business, agency]
        credits:
          type: integer
...
    AdminActivityItem:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
        module:
          type: string
        projectName:
          type: string
        userEmail:
          type: string
        userName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - module
        - createdAt
...